requirements.txt
anyio
asgiref
click
contextlib2
contextvars
dataclasses
fastapi==0.78.0
gunicorn==20.1.0
h11
idna
immutables
importlib-metadata
numpy==1.19.4
opencv-python==4.5.2.54
Pillow
pydantic
python-multipart
six
sniffio
starlette
typing-extensions
uvicorn==0.16.0
zipp
